refactor(apps/web/인게임): useChatStore와 useUserStore를 stores 폴더 아래로 이전
