+++ apps/web/src/puzzlepop-client/stores/useChatStore.ts
@@ -1,7 +1,7 @@
 import { create } from "zustand";
-import { socketStaticStore } from "./socketStaticStore";
-import { ChatData } from "./types/chat";
-import { Player } from "./types/base";
+import { socketStaticStore } from "../socketStaticStore";
+import { ChatData } from "../types/chat";
+import { Player } from "../types/base";
 
 const SYSTEM_PREFIX = "[__SYSTEM__]";
